+++ packages/tdesign-react-aigc/site/test-coverage.js
@@ -1,0 +1,446 @@
+module.exports = {
+  "Util": {
+    "statements": "58.47%",
+    "branches": "47.92%",
+    "functions": "63.49%",
+    "lines": "60.29%"
+  },
+  "affix": {
+    "statements": "84.84%",
+    "branches": "61.29%",
+    "functions": "87.5%",
+    "lines": "85.93%"
+  },
+  "alert": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "anchor": {
+    "statements": "93.85%",
+    "branches": "69.56%",
+    "functions": "88%",
+    "lines": "98.05%"
+  },
+  "autoComplete": {
+    "statements": "96.17%",
+    "branches": "90.9%",
+    "functions": "97.05%",
+    "lines": "97.95%"
+  },
+  "avatar": {
+    "statements": "92.64%",
+    "branches": "86.48%",
+    "functions": "75%",
+    "lines": "92.64%"
+  },
+  "backTop": {
+    "statements": "78.68%",
+    "branches": "53.84%",
+    "functions": "83.33%",
+    "lines": "83.92%"
+  },
+  "badge": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "breadcrumb": {
+    "statements": "84.31%",
+    "branches": "53.12%",
+    "functions": "85.71%",
+    "lines": "89.58%"
+  },
+  "button": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "calendar": {
+    "statements": "78.07%",
+    "branches": "53.24%",
+    "functions": "72%",
+    "lines": "80.86%"
+  },
+  "card": {
+    "statements": "100%",
+    "branches": "84.61%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "cascader": {
+    "statements": "93.12%",
+    "branches": "75.8%",
+    "functions": "90.62%",
+    "lines": "94.21%"
+  },
+  "checkbox": {
+    "statements": "90.27%",
+    "branches": "83.01%",
+    "functions": "100%",
+    "lines": "91.3%"
+  },
+  "collapse": {
+    "statements": "96.15%",
+    "branches": "78.94%",
+    "functions": "94.11%",
+    "lines": "96.1%"
+  },
+  "colorPicker": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "comment": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "common": {
+    "statements": "94.33%",
+    "branches": "84.61%",
+    "functions": "100%",
+    "lines": "97.95%"
+  },
+  "configProvider": {
+    "statements": "70.58%",
+    "branches": "66.66%",
+    "functions": "25%",
+    "lines": "68.75%"
+  },
+  "datePicker": {
+    "statements": "59.16%",
+    "branches": "41.42%",
+    "functions": "58.9%",
+    "lines": "62.07%"
+  },
+  "descriptions": {
+    "statements": "98.82%",
+    "branches": "100%",
+    "functions": "95.45%",
+    "lines": "100%"
+  },
+  "dialog": {
+    "statements": "83.53%",
+    "branches": "71.92%",
+    "functions": "79.06%",
+    "lines": "85.62%"
+  },
+  "divider": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "drawer": {
+    "statements": "86.44%",
+    "branches": "84.48%",
+    "functions": "61.53%",
+    "lines": "89.09%"
+  },
+  "dropdown": {
+    "statements": "89.28%",
+    "branches": "58.69%",
+    "functions": "80%",
+    "lines": "92.59%"
+  },
+  "empty": {
+    "statements": "84.37%",
+    "branches": "63.63%",
+    "functions": "100%",
+    "lines": "84.37%"
+  },
+  "form": {
+    "statements": "83.5%",
+    "branches": "70.73%",
+    "functions": "81.51%",
+    "lines": "87.17%"
+  },
+  "grid": {
+    "statements": "84.21%",
+    "branches": "74.24%",
+    "functions": "90%",
+    "lines": "84.21%"
+  },
+  "guide": {
+    "statements": "99.32%",
+    "branches": "92.85%",
+    "functions": "100%",
+    "lines": "99.31%"
+  },
+  "hooks": {
+    "statements": "61.17%",
+    "branches": "49.41%",
+    "functions": "68.86%",
+    "lines": "62.4%"
+  },
+  "image": {
+    "statements": "88.88%",
+    "branches": "82.53%",
+    "functions": "80%",
+    "lines": "91.86%"
+  },
+  "imageViewer": {
+    "statements": "65.28%",
+    "branches": "76.54%",
+    "functions": "65.11%",
+    "lines": "65.59%"
+  },
+  "input": {
+    "statements": "93.9%",
+    "branches": "92.3%",
+    "functions": "89.47%",
+    "lines": "94.19%"
+  },
+  "inputAdornment": {
+    "statements": "86.95%",
+    "branches": "54.54%",
+    "functions": "100%",
+    "lines": "90.47%"
+  },
+  "inputNumber": {
+    "statements": "76.74%",
+    "branches": "59.74%",
+    "functions": "78.94%",
+    "lines": "80.16%"
+  },
+  "layout": {
+    "statements": "91.48%",
+    "branches": "41.66%",
+    "functions": "85.71%",
+    "lines": "91.48%"
+  },
+  "link": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "list": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "loading": {
+    "statements": "86.07%",
+    "branches": "65%",
+    "functions": "78.57%",
+    "lines": "89.33%"
+  },
+  "locale": {
+    "statements": "73.07%",
+    "branches": "72.22%",
+    "functions": "83.33%",
+    "lines": "73.91%"
+  },
+  "menu": {
+    "statements": "85.44%",
+    "branches": "69.23%",
+    "functions": "81.48%",
+    "lines": "90.51%"
+  },
+  "message": {
+    "statements": "88.96%",
+    "branches": "86.66%",
+    "functions": "70.45%",
+    "lines": "94.28%"
+  },
+  "notification": {
+    "statements": "89.24%",
+    "branches": "75%",
+    "functions": "86.95%",
+    "lines": "92.59%"
+  },
+  "pagination": {
+    "statements": "93.82%",
+    "branches": "76.08%",
+    "functions": "93.75%",
+    "lines": "94.87%"
+  },
+  "popconfirm": {
+    "statements": "76.92%",
+    "branches": "60%",
+    "functions": "81.81%",
+    "lines": "76.92%"
+  },
+  "popup": {
+    "statements": "48.38%",
+    "branches": "44.77%",
+    "functions": "45.23%",
+    "lines": "46.52%"
+  },
+  "progress": {
+    "statements": "89.23%",
+    "branches": "65.71%",
+    "functions": "100%",
+    "lines": "89.23%"
+  },
+  "radio": {
+    "statements": "83.58%",
+    "branches": "47.36%",
+    "functions": "92.85%",
+    "lines": "83.58%"
+  },
+  "rangeInput": {
+    "statements": "75.32%",
+    "branches": "62.79%",
+    "functions": "51.85%",
+    "lines": "75%"
+  },
+  "rate": {
+    "statements": "96.36%",
+    "branches": "80.76%",
+    "functions": "100%",
+    "lines": "96.36%"
+  },
+  "select": {
+    "statements": "100%",
+    "branches": "100%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "selectInput": {
+    "statements": "99%",
+    "branches": "94.11%",
+    "functions": "100%",
+    "lines": "100%"
+  },
+  "skeleton": {
+    "statements": "77.35%",
+    "branches": "43.47%",
+    "functions": "83.33%",
+    "lines": "78.84%"
+  },
+  "slider": {
+    "statements": "89.47%",
+    "branches": "68.85%",
+    "functions": "92.85%",
+    "lines": "91.2%"
+  },
+  "space": {
+    "statements": "87.75%",
+    "branches": "84.37%",
+    "functions": "100%",
+    "lines": "87.75%"
+  },
+  "statistic": {
+    "statements": "84.44%",
+    "branches": "85.71%",
+    "functions": "72.72%",
+    "lines": "85.71%"
+  },
+  "steps": {
+    "statements": "87.8%",
+    "branches": "66.07%",
+    "functions": "100%",
+    "lines": "87.8%"
+  },
+  "swiper": {
+    "statements": "71.93%",
+    "branches": "43.28%",
+    "functions": "85.71%",
+    "lines": "71.35%"
+  },
+  "switch": {
+    "statements": "96.55%",
+    "branches": "92%",
+    "functions": "100%",
+    "lines": "96.55%"
+  },
+  "table": {
+    "statements": "48.36%",
+    "branches": "33.74%",
+    "functions": "45.91%",
+    "lines": "49.56%"
+  },
+  "tabs": {
+    "statements": "89.79%",
+    "branches": "77.27%",
+    "functions": "88%",
+    "lines": "90.86%"
+  },
+  "tag": {
+    "statements": "56.25%",
+    "branches": "48.21%",
+    "functions": "47.05%",
+    "lines": "55.55%"
+  },
+  "tagInput": {
+    "statements": "85.11%",
+    "branches": "82.89%",
+    "functions": "84.21%",
+    "lines": "87.26%"
+  },
+  "textarea": {
+    "statements": "82.89%",
+    "branches": "62.22%",
+    "functions": "80.95%",
+    "lines": "86.76%"
+  },
+  "timePicker": {
+    "statements": "82.88%",
+    "branches": "73.68%",
+    "functions": "86.36%",
+    "lines": "83.01%"
+  },
+  "timeline": {
+    "statements": "96.87%",
+    "branches": "88.13%",
+    "functions": "90.9%",
+    "lines": "96.77%"
+  },
+  "tooltip": {
+    "statements": "90.74%",
+    "branches": "64.7%",
+    "functions": "75%",
+    "lines": "90.56%"
+  },
+  "transfer": {
+    "statements": "86.27%",
+    "branches": "67.61%",
+    "functions": "84.28%",
+    "lines": "87.97%"
+  },
+  "tree": {
+    "statements": "86.22%",
+    "branches": "70.64%",
+    "functions": "84.9%",
+    "lines": "88.33%"
+  },
+  "treeSelect": {
+    "statements": "95.45%",
+    "branches": "82.35%",
+    "functions": "97.56%",
+    "lines": "97.2%"
+  },
+  "typography": {
+    "statements": "95.52%",
+    "branches": "76.31%",
+    "functions": "81.81%",
+    "lines": "98.43%"
+  },
+  "upload": {
+    "statements": "96.77%",
+    "branches": "95.65%",
+    "functions": "88.88%",
+    "lines": "100%"
+  },
+  "watermark": {
+    "statements": "95.77%",
+    "branches": "79.41%",
+    "functions": "100%",
+    "lines": "98.5%"
+  },
+  "utils": {
+    "statements": "75.43%",
+    "branches": "73.68%",
+    "functions": "83.33%",
+    "lines": "74.54%"
+  }
+};